画面を簡略化
