+++ deep-links/DynamicLinksSandbox/src/navigation/RootStackNav.tsx
@@ -1,13 +1,13 @@
 import {createStackNavigator} from '@react-navigation/stack';
 import React from 'react';
-import {Home, Instructions} from 'screens';
+import {Home, DeepLink} from 'screens';
 
 const nav = createStackNavigator();
 export const RootStackNav: React.FC = () => {
   return (
     <nav.Navigator initialRouteName={Home.name}>
       <nav.Screen name="Home" component={Home} options={{headerShown: false}} />
-      <nav.Screen name="Instructions" component={Instructions} />
+      <nav.Screen name="DeepLink" component={DeepLink} />
     </nav.Navigator>
   );
 };
