+++ deep-links/DynamicLinksSandbox/src/screens/deeplink/index.ts
@@ -1,0 +1,1 @@
+export * from './DeepLink';
