+++ deep-links/DynamicLinksSandbox/src/screens/index.ts
@@ -1,2 +1,2 @@
 export * from './home';
-export * from './instructions';
+export * from './deeplink';
