--- deep-links/DynamicLinksSandbox/src/screens/instructions/index.ts
@@ -1,1 +1,0 @@
-export * from './Instructions';
